docs/bmc-cases/bmc-test-framework/pyproject.toml
[project]
name = "bmc-web-test"
version = "2.0.0"
description = "BMC Web UI Automation Tests (Playwright + pytest + Allure)"
requires-python = ">=3.10"
dependencies = [
    "playwright>=1.40",
    "pytest>=8.0",
    "allure-pytest>=2.13",
    "pyyaml>=6.0",
]

[tool.setuptools.packages.find]
include = ["config*", "core*", "pages*"]
exclude = ["tests*", "reports*", "test_data*"]

[tool.pytest.ini_options]
testpaths = ["tests"]
addopts = ["--alluredir=reports/allure-results", "--tb=short", "-v"]
markers = [
    "smoke: quick smoke test suite",
    "regression: full regression suite",
]
